update output overlay
