+++ prj_output/Makefile
@@ -1,0 +1,4 @@
+make:
+	g++ output.cpp -o output.out `pkg-config --cflags --libs gtk+-3.0`
+clean:
+	rm -rf output.out
