+++ prj_output/do.sh
@@ -1,0 +1,2 @@
+make
+./output.out
