+++ prj_output/readme
@@ -1,3 +1,3 @@
 sudo apt-get install gtk+3.0
 sudo apt-get install libgtk-3-dev
-g++ gtk_test.cpp -o output `pkg-config --cflags --libs gtk+-3.0`
+g++ gtk.cpp -o output `pkg-config --cflags --libs gtk+-3.0`
